+++ common/src/main/java/com/samsthenerd/hexgloop/items/ItemDyeableFidget.java
@@ -1,0 +1,9 @@
+package com.samsthenerd.hexgloop.items;
+
+import net.minecraft.item.DyeableItem;
+
+public class ItemDyeableFidget extends ItemFidget implements DyeableItem{
+    public ItemDyeableFidget(Settings settings, FidgetSettings fidgetSettings){
+        super(settings, fidgetSettings);
+    }
+}